maturita-2021/requirements.txt
tk
pillow
numpy